+++ packages/thumbprint-react/components/Dropdown/test.tsx
@@ -99,6 +99,12 @@ test('adds `dataTest` prop', () => {
     expect(wrapperA).toMatchSnapshot();
 });
 
+test('adds `dataTestId` prop', () => {
+    const wrapperA = mount(<Dropdown dataTestId="Duck" onChange={noop} value="goose" />);
+    expect(wrapperA.find('select').prop('data-testid')).toEqual('Duck');
+    expect(wrapperA).toMatchSnapshot();
+});
+
 test('adds `name` HTML attribute', () => {
     const wrapperA = mount(<Dropdown name="duck" onChange={noop} value="goose" />);
     expect(wrapperA.find('select').prop('name')).toEqual('duck');